Only install tools that don't exists

+++ setup.sh
@@ -1,9 +1,11 @@
 #!/usr/bin/bash
 
 # install mise
-curl https://mise.run | sh
-echo 'eval "$(~/.local/bin/mise activate bash)"' >>~/.bashrc
-eval "$(~/.local/bin/mise activate bash)"
+if ! command -v "mise" >/dev/null 2>&1; then
+  curl https://mise.run | sh
+  echo 'eval "$(~/.local/bin/mise activate bash)"' >>~/.bashrc
+  eval "$(~/.local/bin/mise activate bash)"
+fi
 
 # FIX: install tmux (using mise), not working
 # mise plugin add tmux
@@ -11,17 +13,21 @@ eval "$(~/.local/bin/mise activate bash)"
 # mise use -g tmux@latest
 
 # install neovim
-mise plugin add neovim
-mise install neovim@latest
-mise use -g neovim@latest
+if ! command -v "nvim" >/dev/null 2>&1; then
+  mise plugin add neovim
+  mise install neovim@latest
+  mise use -g neovim@latest
+fi
 
 # install lazyvim
-mv ~/.config/nvim{,.bak}
-git clone https://github.com/LazyVim/starter ~/.config/nvim
-rm -rf ~/.config/nvim/.git
+if [[ ! -f ~/.config/nvim/lazyvim.json ]]; then
+  mv ~/.config/nvim{,.bak}
+  git clone https://github.com/LazyVim/starter ~/.config/nvim
+  rm -rf ~/.config/nvim/.git
+fi
 
 # install miniconda
-if ! command -v "conda" >/dev/null 2>&1; then
+if [[ ! -f ~/miniconda3/bin/activate ]]; then
   mkdir -p ~/miniconda3
   wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O ~/miniconda3/miniconda.sh
   bash ~/miniconda3/miniconda.sh -b -u -p ~/miniconda3
@@ -32,18 +38,22 @@ if ! command -v "conda" >/dev/null 2>&1; then
 
   conda tos accept --override-channels --channel https://repo.anaconda.com/pkgs/main
   conda tos accept --override-channels --channel https://repo.anaconda.com/pkgs/r
+
+  # install tmux (using conda)
+  conda install -y conda-forge::tmux
 fi
 
-# install tmux (using conda)
-conda install -y conda-forge::tmux
-
 # install atuin.sh
-bash <(curl --proto '=https' --tlsv1.2 -sSf https://setup.atuin.sh)
+if ! command -v "atuin" >/dev/null 2>&1; then
+  bash <(curl --proto '=https' --tlsv1.2 -sSf https://setup.atuin.sh)
+fi
 
 # install github-cli (using mise)
-mise plugin add github-cli
-mise install github-cli@latest
-mise use -g github-cli@latest
+if ! command -v "gh" >/dev/null 2>&1; then
+  mise plugin add github-cli
+  mise install github-cli@latest
+  mise use -g github-cli@latest
+fi
 
 # set zsh as default
 sudo chsh "$(id -un)" --shell "/usr/bin/zsh"
@@ -52,11 +62,6 @@ sudo chsh "$(id -un)" --shell "/usr/bin/zsh"
 if [[ ! -f ~/antigen.zsh ]]; then
   curl -L git.io/antigen >~/antigen.zsh
 fi
-
-# install spaceship prompt
-# tmp="$(mktemp)"
-# curl -fsSL https://starship.rs/install.sh -o "$tmp" && sh "$tmp" --yes
-# rm -f "$tmp"
 
 # setup dotfiles
 ./copy-dotfiles.sh
